switch all .S files to unified asm, and use a macro to setup compiler to help porting to other compilers. Also some minor tweaks:
* switch some code to use more recent helper methods (e.g. busy_wait_at_least_n_cycles)
* add documentation to host divider header (I had this ages ago and never promoted)
* Some compiler warnings

+++ test/pico_float_test/pico_float_test.c
@@ -518,7 +518,7 @@ int main() {
     }
     for(float x = 4294967296.f * 4294967296.f * 2.f; x>=0.5f; x/=2.f) {
         printf("f2i64 %f->%lld\n", x, (int64_t)x);
-        if (x >= INT64_MAX) {
+        if ((double)x >= (double)INT64_MAX) {
             // seems like there is a bug in the gcc version!
             assert(__aeabi_f2lz(x) == INT64_MAX);
         } else {
